[ex1] Merge sort + Counting algorithm
This is canonical version of this invertions counting algorithm.

It's running time on given example is something about 6.7 seconds on my
Core i5. Which is 550+ times faster than brute_force method.

PS. Also there is a @debug decorator which is a bit talkie, so
merge_and_count.py should be run with -O or -OO.